+++ CloudShell.Providers.Applications/ApplicationProviderExtension.cs
@@ -61,6 +61,7 @@ public sealed class ApplicationProviderExtension : ICloudShellExtension
         builder.Services.TryAddSingleton<ApplicationResourceDeclarationOperations>();
         builder.Services.TryAddSingleton<ApplicationResourceTemplateOperations>();
         builder.Services.TryAddSingleton<ApplicationHostScopedResourceCleanupProvider>();
+        builder.Services.TryAddSingleton<ApplicationResourceSettingsProvider>();
         builder.Services.TryAddSingleton<ApplicationContainerHistoryService>();
         builder.Services.TryAddSingleton<SqlServerDatabaseInspectionService>();
         builder.Services.TryAddSingleton<SqlServerCredentialResolutionService>();
@@ -102,9 +103,9 @@ public sealed class ApplicationProviderExtension : ICloudShellExtension
         builder.Services.AddSingleton<IResourceMonitoringProvider>(
             serviceProvider => serviceProvider.GetRequiredService<ApplicationResourceService>());
         builder.Services.AddSingleton<IResourceAppSettingConfigurationProvider>(
-            serviceProvider => serviceProvider.GetRequiredService<ApplicationResourceService>());
+            serviceProvider => serviceProvider.GetRequiredService<ApplicationResourceSettingsProvider>());
         builder.Services.AddSingleton<IResourceEnvironmentVariableConfigurationProvider>(
-            serviceProvider => serviceProvider.GetRequiredService<ApplicationResourceService>());
+            serviceProvider => serviceProvider.GetRequiredService<ApplicationResourceSettingsProvider>());
         builder.Services.AddSingleton<IHostScopedResourceCleanupProvider>(
             serviceProvider => serviceProvider.GetRequiredService<ApplicationHostScopedResourceCleanupProvider>());
         builder.Services.AddSingleton<IResourceOrchestrationDescriptorProvider>(